requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.7
coverage==7.3.0
cryptography==41.0.3
dictdiffer==0.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
iso8601==1.1.0
mangum==0.17.0
more-itertools==10.1.0
packaging==23.1
passlib==1.7.4
pluggy==0.13.1
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pypika-tortoise==0.1.6
pytest==5.4.3
pytest-cov==2.12.1
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tomlkit==0.12.1
tortoise-orm==0.20.0
typing_extensions==4.7.1
uvicorn==0.23.2
wcwidth==0.2.6
